api/requirements.txt
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.9.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.5
googletrans==2.3.0
gunicorn==19.9.0
idna==2.7
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
